Cleanup ServiceRequestBase.buildEwsHttpWebRequest()
The old implementation seemed to be a 1:1 port of C# code, that took into
account all kinds of crap that only happens in C#'s HttpWebRequest (see [1],
which was also mentioned in the original code), like I/O delays on the creation
of OutputStreams. The new implementation could still use some work (e.g.,
signing is completely unsupported, the outputstream is cast to a
ByteArrayOutputStream (was already in the original code)), but at least this
code is readable.

[1]: http://www.wintellect.com/blogs/jeffreyr/httpwebrequest-its-request-stream-and-sending-data-in-chunks