Adding a java/olson TimeZone conversion class
This is useful for setting appointments with timezones when you have
the java timezone. Otherwise there is no way to get the proper timezone
as they won’t have periods (even the getServerTimezones doesn’t have
the information needed). I’ve been using this for months (before all
this was on github) with great results